+++ KeeAnywhere/Forms/SettingsForm.Designer.cs
@@ -43,7 +43,7 @@
             this.m_rbStorageLocation_LocalUserSecureStore = new System.Windows.Forms.RadioButton();
             this.lblAccountStorageLocation = new System.Windows.Forms.Label();
             this.m_tabGeneral = new System.Windows.Forms.TabPage();
-            this.m_cbOfflineCache = new System.Windows.Forms.CheckBox();
+            this.m_chkOfflineCache = new System.Windows.Forms.CheckBox();
             this.m_tabAbout = new System.Windows.Forms.TabPage();
             this.m_lblAboutVersion = new System.Windows.Forms.Label();
             this.m_lblAboutExplanation = new System.Windows.Forms.Label();
@@ -61,6 +61,9 @@
             this.m_mnuHelp_Sep2 = new System.Windows.Forms.ToolStripSeparator();
             this.m_mnuHelp_Donate = new System.Windows.Forms.ToolStripMenuItem();
             this.m_pnlFormButtons = new System.Windows.Forms.TableLayoutPanel();
+            this.m_gbOfflineCache = new System.Windows.Forms.GroupBox();
+            this.m_btnClearCache = new System.Windows.Forms.Button();
+            this.m_btnOpenCacheFolder = new System.Windows.Forms.Button();
             this.m_btnHelp = new KeeAnywhere.Forms.DropDownButton();
             this.m_btnAccountAdd = new KeeAnywhere.Forms.DropDownButton();
             ((System.ComponentModel.ISupportInitialize)(this.m_bannerImage)).BeginInit();
@@ -70,6 +73,7 @@
             this.m_tabAbout.SuspendLayout();
             this.m_mnuHelp.SuspendLayout();
             this.m_pnlFormButtons.SuspendLayout();
+            this.m_gbOfflineCache.SuspendLayout();
             this.SuspendLayout();
             // 
             // m_bannerImage
@@ -209,7 +213,7 @@
             // 
             // m_tabGeneral
             // 
-            this.m_tabGeneral.Controls.Add(this.m_cbOfflineCache);
+            this.m_tabGeneral.Controls.Add(this.m_gbOfflineCache);
             this.m_tabGeneral.Location = new System.Drawing.Point(4, 22);
             this.m_tabGeneral.Name = "m_tabGeneral";
             this.m_tabGeneral.Padding = new System.Windows.Forms.Padding(3);
@@ -218,15 +222,16 @@
             this.m_tabGeneral.Text = "General";
             this.m_tabGeneral.UseVisualStyleBackColor = true;
             // 
-            // m_cbOfflineCache
+            // m_chkOfflineCache
             // 
-            this.m_cbOfflineCache.AutoSize = true;
-            this.m_cbOfflineCache.Location = new System.Drawing.Point(7, 7);
-            this.m_cbOfflineCache.Name = "m_cbOfflineCache";
-            this.m_cbOfflineCache.Size = new System.Drawing.Size(189, 17);
-            this.m_cbOfflineCache.TabIndex = 0;
-            this.m_cbOfflineCache.Text = "Cache Databases for offline usage";
-            this.m_cbOfflineCache.UseVisualStyleBackColor = true;
+            this.m_chkOfflineCache.AutoSize = true;
+            this.m_chkOfflineCache.Location = new System.Drawing.Point(6, 19);
+            this.m_chkOfflineCache.Name = "m_chkOfflineCache";
+            this.m_chkOfflineCache.Size = new System.Drawing.Size(189, 17);
+            this.m_chkOfflineCache.TabIndex = 0;
+            this.m_chkOfflineCache.Text = "Cache Databases for offline usage";
+            this.m_chkOfflineCache.UseVisualStyleBackColor = true;
+            this.m_chkOfflineCache.CheckedChanged += new System.EventHandler(this.OnOfflineCacheChanged);
             // 
             // m_tabAbout
             // 
@@ -394,6 +399,40 @@
             this.m_pnlFormButtons.Size = new System.Drawing.Size(604, 29);
             this.m_pnlFormButtons.TabIndex = 9;
             // 
+            // m_gbOfflineCache
+            // 
+            this.m_gbOfflineCache.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.m_gbOfflineCache.Controls.Add(this.m_btnOpenCacheFolder);
+            this.m_gbOfflineCache.Controls.Add(this.m_btnClearCache);
+            this.m_gbOfflineCache.Controls.Add(this.m_chkOfflineCache);
+            this.m_gbOfflineCache.Location = new System.Drawing.Point(6, 6);
+            this.m_gbOfflineCache.Name = "m_gbOfflineCache";
+            this.m_gbOfflineCache.Size = new System.Drawing.Size(578, 84);
+            this.m_gbOfflineCache.TabIndex = 1;
+            this.m_gbOfflineCache.TabStop = false;
+            this.m_gbOfflineCache.Text = "Offline Cache";
+            // 
+            // m_btnClearCache
+            // 
+            this.m_btnClearCache.Location = new System.Drawing.Point(6, 43);
+            this.m_btnClearCache.Name = "m_btnClearCache";
+            this.m_btnClearCache.Size = new System.Drawing.Size(75, 23);
+            this.m_btnClearCache.TabIndex = 1;
+            this.m_btnClearCache.Text = "Clear Cache";
+            this.m_btnClearCache.UseVisualStyleBackColor = true;
+            this.m_btnClearCache.Click += new System.EventHandler(this.OnClearCache);
+            // 
+            // m_btnOpenCacheFolder
+            // 
+            this.m_btnOpenCacheFolder.Location = new System.Drawing.Point(88, 43);
+            this.m_btnOpenCacheFolder.Name = "m_btnOpenCacheFolder";
+            this.m_btnOpenCacheFolder.Size = new System.Drawing.Size(182, 23);
+            this.m_btnOpenCacheFolder.TabIndex = 2;
+            this.m_btnOpenCacheFolder.Text = "Open Cache Folder in Explorer";
+            this.m_btnOpenCacheFolder.UseVisualStyleBackColor = true;
+            this.m_btnOpenCacheFolder.Click += new System.EventHandler(this.OnOpenCacheFolder);
+            // 
             // m_btnHelp
             // 
             this.m_btnHelp.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
@@ -440,10 +479,11 @@
             this.m_tabAccounts.ResumeLayout(false);
             this.m_tabAccounts.PerformLayout();
             this.m_tabGeneral.ResumeLayout(false);
-            this.m_tabGeneral.PerformLayout();
             this.m_tabAbout.ResumeLayout(false);
             this.m_mnuHelp.ResumeLayout(false);
             this.m_pnlFormButtons.ResumeLayout(false);
+            this.m_gbOfflineCache.ResumeLayout(false);
+            this.m_gbOfflineCache.PerformLayout();
             this.ResumeLayout(false);
 
         }
@@ -457,7 +497,7 @@
         private System.Windows.Forms.TabPage m_tabAbout;
         private System.Windows.Forms.Button m_btnOK;
         private System.Windows.Forms.Button m_btnCancel;
-        private System.Windows.Forms.CheckBox m_cbOfflineCache;
+        private System.Windows.Forms.CheckBox m_chkOfflineCache;
         private System.Windows.Forms.Label m_lblAboutExplanation;
         private System.Windows.Forms.Label m_lblAboutHeader;
         private System.Windows.Forms.RadioButton m_rbStorageLocation_Disk;
@@ -483,5 +523,8 @@
         private System.Windows.Forms.ToolStripSeparator m_mnuHelp_Sep2;
         private System.Windows.Forms.ToolStripMenuItem m_mnuHelp_Donate;
         private System.Windows.Forms.TableLayoutPanel m_pnlFormButtons;
+        private System.Windows.Forms.GroupBox m_gbOfflineCache;
+        private System.Windows.Forms.Button m_btnOpenCacheFolder;
+        private System.Windows.Forms.Button m_btnClearCache;
     }
 }