+++ KeeAnywhere/StorageProviders/Box/BoxStorageProvider.cs
@@ -32,7 +32,7 @@ namespace KeeAnywhere.StorageProviders.Box
             return stream;
         }
 
-        public async Task<bool> Save(Stream stream, string path)
+        public async Task Save(Stream stream, string path)
         {
             var api = await GetApi();
             BoxFile item;
@@ -62,7 +62,8 @@ namespace KeeAnywhere.StorageProviders.Box
 
             }
 
-            return item != null;
+            if (item == null)
+                throw new InvalidOperationException("Save to Box failed.");
         }